refactor: create common component Games
