--- src/pages/games/audioChallenge/dataGame.ts
@@ -1,14 +1,0 @@
-import { IDataGame } from '../interfaces/IDataGame';
-
-const dataGame: IDataGame = {
-  name: 'Аудиовызов',
-  goal: 'Цель - выбрать перевод слова по звучащему произношению.',
-  description: `В процессе игры звучит произношение слова на английском языке, нужно выбрать перевод слова из пяти предложенных вариантов ответа.
-          Слова можно угадывать, выбирая их как кликами мышкой, так и нажатием кнопок клавиатуры от 1 до 5.
-          Переход к следующему вопросу происходит как при клике по стрелке, так и нажатием клавиши Enter.
-          Воспроизвести произношение слова можно как при клике по кнопке "Аудио", так и нажатием клавиши R.`,
-  bg: '#f3831a',
-  amountWords: 50,
-};
-
-export default dataGame;
