+++ src/pages/games/helpers/gamesData.ts
@@ -7,13 +7,15 @@ export const sprintData = {
   amountWords: 600,
 };
 
-export const dataGame = {
-  name: 'Спринт',
-  goal: 'Цель: Повторить слова и их переводы',
-  description:
-    'В этой игре вам предстоит сравнивать слово на английском с его переводом на русском языке. Если слово и перевод соответствуют друг другу - нажмите стрелочку вправо либо на кнопку "верно", если же нет - нажмите на стрелочку влево, либо на кнопку "неверно". Время на игру: 60 секунд',
-  bg: 'blue',
-  amountWords: 600,
+export const audioChallengeData = {
+  name: 'Аудиовызов',
+  goal: 'Цель - выбрать перевод слова по звучащему произношению.',
+  description: `В процессе игры звучит произношение слова на английском языке, нужно выбрать перевод слова из пяти предложенных вариантов ответа.
+          Слова можно угадывать, выбирая их как кликами мышкой, так и нажатием кнопок клавиатуры от 1 до 5.
+          Переход к следующему вопросу происходит как при клике по стрелке, так и нажатием клавиши Enter.
+          Воспроизвести произношение слова можно как при клике по кнопке "Аудио", так и нажатием клавиши R.`,
+  bg: '#f3831a',
+  amountWords: 50,
 };
 
 export const wordConstructorData = {
